+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Updated python-lib to v3.0.3
 - Updated mysql-lib to v5.3.4
 
+### Fixed
+- run_program: Added check to determine server_id format based on MySQL version running.
+
 ### Changed
 - Set simplejson to 3.12.0 for Python 3.
 - Set chardet to 3.0.4 for Python 2.